+++ shared/src/test/scala/FetchTests.scala
@@ -99,90 +99,6 @@ object TestHelper {
     rs.map((round: Round) => requestBatches(round.request)).toList.sum
 }
 
-class FetchSyntaxTests extends AsyncFreeSpec with Matchers {
-  import fetch.syntax._
-  import TestHelper._
-
-  val ME = implicitly[FetchMonadError[Future]]
-
-  implicit override def executionContext = ExecutionContext.Implicits.global
-
-  "Cartesian syntax is implicitly concurrent" in {
-    import cats.syntax.cartesian._
-
-    val fetch: Fetch[(Int, List[Int])] = (one(1) |@| many(3)).tupled
-
-    val fut = Fetch.runEnv[Future](fetch)
-
-    fut.map(env => {
-      env.rounds.size shouldEqual 1
-    })
-  }
-
-  "Apply syntax is implicitly concurrent" in {
-    import cats.syntax.apply._
-
-    val fetch: Fetch[Int] = Fetch.pure((x: Int, y: Int) => x + y).ap2(one(1), one(2))
-
-    val fut = Fetch.runEnv[Future](fetch)
-
-    fut.map(env => {
-      val rounds = env.rounds
-
-      rounds.size shouldEqual 1
-      totalBatches(rounds) shouldEqual 1
-      totalFetched(rounds) shouldEqual 2
-    })
-  }
-
-  "`fetch` syntax allows lifting of any value to the context of a fetch" in {
-    Fetch.pure(42) shouldEqual 42.fetch
-  }
-
-  "`fetch` syntax allows lifting of any `Throwable` as a failure on a fetch" in {
-    case object Ex extends RuntimeException
-
-    val fut1 = Fetch.run[Future](Fetch.error(Ex): Fetch[Int])
-    val fut2 = Fetch.run[Future](Ex.fetch: Fetch[Int])
-
-    val e1 = ME.handleErrorWith(fut1)(err => Future.successful(42))
-    val e2 = ME.handleErrorWith(fut2)(err => Future.successful(42))
-
-    ME.map2(e1, e2)(_ shouldEqual _)
-  }
-
-  "`join` syntax is equivalent to `Fetch#join`" in {
-    val join1 = Fetch.join(one(1), many(3))
-    val join2 = one(1).join(many(3))
-
-    ME.map2(Fetch.run[Future](join1), Fetch.run[Future](join2))(_ shouldEqual _)
-  }
-
-  "`runF` syntax is equivalent to `Fetch#runFetch`" in {
-
-    val rf1 = Fetch.runFetch[Future](1.fetch)
-    val rf2 = 1.fetch.runF[Future]
-
-    ME.map2(rf1, rf2)(_ shouldEqual _)
-  }
-
-  "`runE` syntax is equivalent to `Fetch#runEnv`" in {
-
-    val rf1 = Fetch.runEnv[Future](1.fetch)
-    val rf2 = 1.fetch.runE[Future]
-
-    ME.map2(rf1, rf2)(_ shouldEqual _)
-  }
-
-  "`runA` syntax is equivalent to `Fetch#run`" in {
-
-    val rf1 = Fetch.run[Future](1.fetch)
-    val rf2 = 1.fetch.runA[Future]
-
-    ME.map2(rf1, rf2)(_ shouldEqual _)
-  }
-}
-
 class FetchTests extends AsyncFreeSpec with Matchers {
   import TestHelper._
 
@@ -806,237 +722,6 @@ class FetchTests extends AsyncFreeSpec with Matchers {
     Fetch.runFetch[Future](fetch).map {
       case (env, res) =>
         res shouldEqual (1, 5)
-        totalFetched(env.rounds) shouldEqual 2
-        env.rounds.size shouldEqual 1
-    }
-  }
-}
-
-class FetchReportingTests extends AsyncFreeSpec with Matchers {
-  import TestHelper._
-
-  val ME = implicitly[FetchMonadError[Future]]
-
-  implicit override def executionContext = ExecutionContext.Implicits.global
-
-  "Plain values have no rounds of execution" in {
-    val fetch: Fetch[Int] = Fetch.pure(42)
-    Fetch.runEnv[Future](fetch).map(_.rounds.size shouldEqual 0)
-  }
-
-  "Single fetches are executed in one round" in {
-    val fetch = one(1)
-    Fetch.runEnv[Future](fetch).map(_.rounds.size shouldEqual 1)
-  }
-
-  "Single fetches are executed in one round per binding in a for comprehension" in {
-    val fetch = for {
-      o <- one(1)
-      t <- one(2)
-    } yield (o, t)
-
-    Fetch.runEnv[Future](fetch).map(_.rounds.size shouldEqual 2)
-  }
-
-  "Single fetches for different data sources are executed in multiple rounds if they are in a for comprehension" in {
-    val fetch: Fetch[(Int, List[Int])] = for {
-      o <- one(1)
-      m <- many(3)
-    } yield (o, m)
-
-    Fetch.runEnv[Future](fetch).map(_.rounds.size shouldEqual 2)
-  }
-
-  "Single fetches combined with cartesian are run in one round" in {
-    import cats.syntax.cartesian._
-
-    val fetch: Fetch[(Int, List[Int])] = (one(1) |@| many(3)).tupled
-    val fut                            = Fetch.runEnv[Future](fetch)
-
-    fut.map(_.rounds.size shouldEqual 1)
-  }
-
-  "Single fetches combined with traverse are run in one round" in {
-    import cats.syntax.traverse._
-
-    val fetch: Fetch[List[Int]] = for {
-      manies <- many(3)
-      ones   <- manies.traverse(one)
-    } yield ones
-
-    val fut = Fetch.runEnv[Future](fetch)
-    fut.map(_.rounds.size shouldEqual 2)
-  }
-
-  "The product of two fetches from the same data source implies batching" in {
-    val fetch: Fetch[(Int, Int)] = Fetch.join(one(1), one(3))
-
-    Fetch
-      .runEnv[Future](fetch)
-      .map(env => {
-        env.rounds.size shouldEqual 1
-      })
-  }
-
-  "The product of concurrent fetches of the same type implies everything fetched in batches" in {
-    val fetch = Fetch.join(
-      Fetch.join(
-        for {
-          a <- one(1)
-          b <- one(2)
-          c <- one(3)
-        } yield c,
-        for {
-          a <- one(2)
-          m <- many(4)
-          c <- one(3)
-        } yield c
-      ),
-      one(3)
-    )
-
-    Fetch
-      .runEnv[Future](fetch)
-      .map(env => {
-        env.rounds.size shouldEqual 2
-      })
-  }
-
-  "Every level of sequenced concurrent of concurrent fetches is batched" in {
-    val fetch = Fetch.join(
-      Fetch.join(
-        for {
-          a <- Fetch.sequence(List(one(2), one(3), one(4)))
-          b <- Fetch.sequence(List(many(0), many(1)))
-          c <- Fetch.sequence(List(one(9), one(10), one(11)))
-        } yield c,
-        for {
-          a <- Fetch.sequence(List(one(5), one(6), one(7)))
-          b <- Fetch.sequence(List(many(2), many(3)))
-          c <- Fetch.sequence(List(one(12), one(13), one(14)))
-        } yield c
-      ),
-      Fetch.sequence(List(one(15), one(16), one(17)))
-    )
-
-    Fetch
-      .runEnv[Future](fetch)
-      .map(env => {
-        env.rounds.size shouldEqual 3
-      })
-  }
-}
-
-class FetchBatchingTests extends AsyncFreeSpec with Matchers {
-  import TestHelper._
-
-  implicit override def executionContext = ExecutionContext.Implicits.global
-
-  case class BatchedDataSeq(id: Int)
-  implicit object MaxBatchSourceSeq extends DataSource[BatchedDataSeq, Int] {
-    override def name = "BatchSourceSeq"
-    override def fetchOne(id: BatchedDataSeq): Query[Option[Int]] = {
-      Query.sync(Option(id.id))
-    }
-    override def fetchMany(ids: NonEmptyList[BatchedDataSeq]): Query[Map[BatchedDataSeq, Int]] =
-      Query.sync(ids.toList.map(one => (one, one.id)).toMap)
-
-    override val maxBatchSize = Some(2)
-
-    override val batchExecution = Sequential
-  }
-
-  case class BatchedDataPar(id: Int)
-  implicit object MaxBatchSourcePar extends DataSource[BatchedDataPar, Int] {
-    override def name = "BatchSourcePar"
-    override def fetchOne(id: BatchedDataPar): Query[Option[Int]] = {
-      Query.sync(Option(id.id))
-    }
-    override def fetchMany(ids: NonEmptyList[BatchedDataPar]): Query[Map[BatchedDataPar, Int]] =
-      Query.sync(ids.toList.map(one => (one, one.id)).toMap)
-
-    override val maxBatchSize = Some(2)
-
-    override val batchExecution = Parallel
-  }
-
-  def fetchBatchedDataSeq(id: Int): Fetch[Int] = Fetch(BatchedDataSeq(id))
-  def fetchBatchedDataPar(id: Int): Fetch[Int] = Fetch(BatchedDataPar(id))
-
-  "A large fetch to a datasource with a maximum batch size is split and executed in sequence" in {
-    val fetch: Fetch[List[Int]] = Fetch.traverse(List.range(1, 6))(fetchBatchedDataSeq)
-    Fetch.runFetch[Future](fetch).map {
-      case (env, res) =>
-        res shouldEqual List(1, 2, 3, 4, 5)
-        totalFetched(env.rounds) shouldEqual 5
-        totalBatches(env.rounds) shouldEqual 2
-        env.rounds.size shouldEqual 3
-    }
-  }
-
-  "A large fetch to a datasource with a maximum batch size is split and executed in parallel" in {
-    val fetch: Fetch[List[Int]] = Fetch.traverse(List.range(1, 6))(fetchBatchedDataPar)
-    Fetch.runFetch[Future](fetch).map {
-      case (env, res) =>
-        res shouldEqual List(1, 2, 3, 4, 5)
-        totalFetched(env.rounds) shouldEqual 5
-        totalBatches(env.rounds) shouldEqual 2
-        env.rounds.size shouldEqual 1
-    }
-  }
-
-  "Fetches to datasources with a maximum batch size should be split and executed in parallel and sequentially" in {
-    val fetch: Fetch[List[Int]] =
-      Fetch.traverse(List.range(1, 6))(fetchBatchedDataPar) *>
-        Fetch.traverse(List.range(1, 6))(fetchBatchedDataSeq)
-
-    Fetch.runFetch[Future](fetch).map {
-      case (env, res) =>
-        res shouldEqual List(1, 2, 3, 4, 5)
-        totalFetched(env.rounds) shouldEqual 5 + 5
-        totalBatches(env.rounds) shouldEqual 2 + 2
-        env.rounds.size shouldEqual 3
-    }
-  }
-
-  "A large (many) fetch to a datasource with a maximum batch size is split and executed in sequence" in {
-    val fetch: Fetch[List[Int]] =
-      Fetch.multiple(BatchedDataSeq(1), BatchedDataSeq(2), BatchedDataSeq(3))
-    Fetch.runFetch[Future](fetch).map {
-      case (env, res) =>
-        res shouldEqual List(1, 2, 3)
-        totalFetched(env.rounds) shouldEqual 3
-        totalBatches(env.rounds) shouldEqual 2 // FetchMany(NEL(1, 2)) and FetchMany(NEL(3))
-        env.rounds.size shouldEqual 2
-    }
-  }
-
-  "A large (many) fetch to a datasource with a maximum batch size is split and executed in parallel" in {
-    val fetch: Fetch[List[Int]] =
-      Fetch.multiple(BatchedDataPar(1), BatchedDataPar(2), BatchedDataPar(3))
-    Fetch.runFetch[Future](fetch).map {
-      case (env, res) =>
-        res shouldEqual List(1, 2, 3)
-        totalFetched(env.rounds) shouldEqual 3
-        totalBatches(env.rounds) shouldEqual 1
-        env.rounds.size shouldEqual 1
-
-    }
-  }
-
-  "Very deep fetches don't overflow stack or heap" in {
-    import cats.syntax.traverse._
-
-    val depth = 20
-    val fetch: Fetch[List[Int]] = (1 to depth).toList
-      .map((x) => (0 until 2).toList.traverse(fetchBatchedData))
-      .foldLeft(
-        Fetch.pure(List.empty[Int])
-      )((acc, f) => acc.flatMap((x) => f))
-
-    Fetch.runFetch[Future](fetch).map {
-      case (env, res) =>
-        res shouldEqual List(0, 1)
         totalFetched(env.rounds) shouldEqual 2
         env.rounds.size shouldEqual 1
     }
